+++ test/test_it.py
@@ -8,7 +8,7 @@ from timezonefinder.timezonefinder import TimezoneFinder
 from tzwhere.tzwhere import tzwhere
 
 # number of points to test (in each test, realistic and random ones)
-N = 1000
+N = 100
 
 # sets if tzwhere should be used with shapely
 SHAPELY = False
@@ -232,6 +232,7 @@ class PackageEqualityTest(unittest.TestCase):
         print_equality_test('random points', list_of_random_points(length=N))
 
     def test_speed(self):
+
         def check_speed_his_algor(list_of_points):
             start_time = datetime.now()
             for point in list_of_points: