+++ timezonefinder/timezonefinder.py
@@ -29,10 +29,11 @@ from .timezone_names import timezone_names
 try:
     import numba
 
-    print(numba.__version__)
+    print('using numba version:', numba.__version__)
 
     print('compiling the helpers ahead of time...')
     # FIXME target architecture is wrong. because of old Numba version?
+    # TODO in this environment numba could not be available
     system("python3 /Users/jannikmi/GitHub/timezonefinder/timezonefinder/helpers_numba.py")
     try:
         from compiled_helpers import coord2int, distance_to_polygon_exact, distance_to_polygon, inside_polygon, \
@@ -201,11 +202,67 @@ class TimezoneFinder:
         except KeyError:
             return
 
-    def id_list(self, amount_polygons, polygon_nr_list):
-        output = empty([amount_polygons], dtype='i2')
-        for i in range(amount_polygons):
-            output[i] = self.id_of(polygon_nr_list[i])
-        return output
+    def compile_id_list(self, polygon_id_list, nr_of_polygons, dont_sort=False):
+        """
+        sorts the polygons_id list from least to most occurrences of the zone ids (->speed up)
+        :param polygon_id_list:
+        :param nr_of_polygons: length of polygon_id_list
+        :param dont_sort: if this is set to True, the sorting algorithms is skipped
+        :return: sorted list of polygon_ids, sorted list of zone_ids, boolean: do all entries belong to the same zone
+        """
+        # print(polygon_id_list)
+        # print(zone_id_list)
+        counted_zones = {}
+        all_zones_occurred_only_once = True
+        zone_id_list = empty([nr_of_polygons], dtype='<u2', )
+
+        pointer_local = 0
+        for polygon_id in polygon_id_list:
+            zone_id = self.id_of(polygon_id)
+            zone_id_list[pointer_local] = zone_id
+            pointer_local += 1
+            try:
+                counted_zones[zone_id] += 1
+                all_zones_occurred_only_once = False
+            except KeyError:
+                counted_zones[zone_id] = 1
+        # print(counted_zones)
+
+        if len(counted_zones) == 1:
+            return polygon_id_list, zone_id_list, True
+
+        if all_zones_occurred_only_once or dont_sort:
+            return polygon_id_list, zone_id_list, False
+
+        counted_zones_sorted = sorted(counted_zones.items(), key=lambda zone: zone[1])
+        # print(counted_zones_sorted)
+
+        sorted_polygon_id_list = empty([nr_of_polygons], dtype='<u2')
+        sorted_zone_id_list = empty([nr_of_polygons], dtype='<u2')
+
+        pointer_output = 0
+        pointer_output2 = 0
+        for zone_id, amount in counted_zones_sorted:
+            # write all polygons from this zone in the new list
+            pointer_local = 0
+            detected_polygons = 0
+            while detected_polygons < amount:
+                if zone_id_list[pointer_local] == zone_id:
+                    # the polygon at the pointer has the wanted zone_id
+                    detected_polygons += 1
+                    sorted_polygon_id_list[pointer_output] = polygon_id_list[pointer_local]
+                    pointer_output += 1
+
+                pointer_local += 1
+
+            for pointer_local in range(amount):
+                sorted_zone_id_list[pointer_output2] = zone_id
+                pointer_output2 += 1
+
+        # print(sorted_polygon_id_list)
+        # print(sorted_zone_id_list)
+
+        return sorted_polygon_id_list, sorted_zone_id_list, False
 
     def closest_timezone_at(self, lng, lat, delta_degree=1, exact_computation=False, return_distances=False,
                             force_evaluation=False):
@@ -261,7 +318,7 @@ class TimezoneFinder:
         lng = radians(lng)
         lat = radians(lat)
 
-        polygon_nrs = []
+        possible_polygons = []
 
         # there are 2 shortcuts per 1 degree lat, so to cover 1 degree two shortcuts (rows) have to be checked
         # the highest shortcut is 0
@@ -278,27 +335,27 @@ class TimezoneFinder:
         for x in range(left, right + 1, 1):
             for y in range(top, bottom + 1, 1):
                 for p in self.polygons_of_shortcut(x, y):
-                    if p not in polygon_nrs:
-                        polygon_nrs.append(p)
+                    if p not in possible_polygons:
+                        possible_polygons.append(p)
 
-        polygons_in_list = len(polygon_nrs)
+        polygons_in_list = len(possible_polygons)
 
         if polygons_in_list == 0:
             return None
 
         # initialize the list of ids
-        ids = self.id_list(polygons_in_list, polygon_nrs)
+        # TODO sorting doesn't give a bonus here?!
+        possible_polygons, ids, zones_are_equal = self.compile_id_list(possible_polygons, polygons_in_list, dont_sort=True)
 
         # if all the polygons in this shortcut belong to the same zone return it
-        same_element = all_the_same(pointer=0, length=polygons_in_list, id_list=ids)
-        if same_element != -1:
+        if zones_are_equal:
             if not (return_distances or force_evaluation):
-                return timezone_names[same_element]
+                return timezone_names[ids[0]]
 
         distances = [None for i in range(polygons_in_list)]
         pointer = 0
         if force_evaluation:
-            for polygon_nr in polygon_nrs:
+            for polygon_nr in possible_polygons:
                 distance = routine(polygon_nr)
                 distances[pointer] = distance
                 if distance < min_distance:
@@ -319,7 +376,7 @@ class TimezoneFinder:
 
                 else:
                     # this polygon has to be checked
-                    distance = routine(polygon_nrs[pointer])
+                    distance = routine(possible_polygons[pointer])
                     distances[pointer] = distance
 
                     already_checked[pointer] = True
@@ -365,27 +422,24 @@ class TimezoneFinder:
             return timezone_names[self.id_of(possible_polygons[0])]
 
         # initialize the list of ids
-        ids = self.id_list(nr_possible_polygons, possible_polygons)
-        # TODO sort ids AND possible_polygons from least to most occurrences (->speed up?!)
-        self.sort_least_occurrences(possible_polygons, ids)
+        # and sort possible_polygons from least to most occurrences of zone_id
+        possible_polygons, ids, zones_are_equal = self.compile_id_list(possible_polygons, nr_possible_polygons)
+        if zones_are_equal:
+            return timezone_names[ids[0]]
 
         # otherwise check if the point is included for all the possible polygons
         for i in range(nr_possible_polygons):
-
-            same_element = all_the_same(pointer=i, length=nr_possible_polygons, id_list=ids)
-            if same_element != -1:
-                return timezone_names[same_element]
 
             polygon_nr = possible_polygons[i]
 
             # get the boundaries of the polygon = (lng_max, lng_min, lat_max, lat_min)
             self.binary_file.seek((self.bound_start_address + 16 * polygon_nr), )
             boundaries = fromfile(self.binary_file, dtype='<i4', count=4)
-            # only run the algorithm if it the point is withing the boundaries
+            # only run the expensive algorithm if the point is withing the boundaries
             if not (x > boundaries[0] or x < boundaries[1] or y > boundaries[2] or y < boundaries[3]):
 
                 outside_all_holes = True
-                # when the point is within a hole of the polygon this timezone doesn't need to be checked
+                # when the point is within a hole of the polygon, this timezone doesn't need to be checked
                 for hole_coordinates in self._holes_of_line(polygon_nr):
                     if inside_polygon(x, y, hole_coordinates):
                         outside_all_holes = False
@@ -394,6 +448,11 @@ class TimezoneFinder:
                 if outside_all_holes:
                     if inside_polygon(x, y, self.coords_of(line=polygon_nr)):
                         return timezone_names[ids[i]]
+
+            # when after the current polygon only polygon from the same zone appear, return this zone
+            same_element = all_the_same(pointer=i + 1, length=nr_possible_polygons, id_list=ids)
+            if same_element != -1:
+                return timezone_names[same_element]
 
         return None